+++ src/index.js
@@ -1,9 +1,6 @@
-import deepAssign from './utils/deep-assign'
-import each from './utils/each'
 import getNode from './utils/get-node'
 import getNodes from './utils/get-nodes'
 import isNodeList from './utils/is-node-list'
 import isNode from './utils/is-node'
-import isObject from './utils/is-object'
 
-export { deepAssign, each, getNode, getNodes, isNodeList, isNode, isObject }
+export { getNode, getNodes, isNodeList, isNode }