+++ src/utils/is-object.js
@@ -1,7 +1,0 @@
-export default function isObject(x) {
-	return (
-		x !== null &&
-		x instanceof Object &&
-		(x.constructor === Object || Object.prototype.toString.call(x) === '[object Object]')
-	)
-}